Add an external pillar.
The pillar is not programatically modifiable. Since we need to set
some pillar values during the bootstrapping process, we need to rely
on an external pillar that we can modify.

The original module requires a '%s' argument to be provided, or the
mysql python module will complain. We added a `minion_id` that can be
used for certain pillar key discovery. However, since we are going to
use this pillar information for global configuration, we can just let
`minion_id` be NULL, and that will be readable by all salt clients.

We also add a Pillar model that will be able to modify these Pillar
values.