Add AutoYaST profile serving
This profile will be read by workers during the bootstrapping process
and when adding new workers to the cluster. It will provide the autoyast
directives for the unattended installation to take place. Among other
installation configuration settings, there is the forwarding of the
subscription details of the controller node to the workers.

Here we take into account that we are using our SUSE Secrets fork of
Docker, that mounts `/run/secrets` on every container. This directory
will include `/etc/SUSEConnect` file and `/etc/zypp/credentials.d` folder
from the host mounted inside the container.

- `/run/secrets/SUSEConnect` includes the SMT if any (the default is SCC
  if not provided). If there is a SMT present, the regcode will be empty.

- If we are using SUSE's Customer Center (SCC), we need a regcode, that
  is obtained using the credentials in
  `/run/secrets/credentials.d/SCCcredentials` file, accessing the SCC API
   and checking for an active subscription. Filtering the CaaSP active
   subscription we can fetch the registration code for the subscription
   and serve it.