Retrieve host from request
Since `Socket.gethostname` can be a purely internal hostname (not
exposed to other machines) and since the customer is in charge of
DNS itself, and since `@controller_node` here is going to be used
to provide a reachable salt master hostname for the rest of the
workers, it's better to just use the same hostname the customer used
when added the `autoyast=http://fancy-host.fancy-domain/autoyast`. In
this case we don't care about the internal name `Socket.gethostname`
but about the external name (the customer already used it on their own
request, so we just retrieve it).