Remove port from puma configuration
We are firing puma up with something along the lines:
`bundle exec rails s -b 0.0.0.0 -p 80 --pid /tmp/puma.pid Puma`

In this case, the port is set through an argument to the rails server
command, but with this change on our puma configuration, we override
it, only honoring the `PORT` envvar, or falling back to 3000
otherwise.

Remove this bit from the Puma configuration, since our manifests
include the port as an argument to the rails server command.