add init for containers
add a wrapper script called init that will check if the database has
been configured and, in that case, it will configure it before starting
the rails server.

This script calls db:migrate because this is meant to be run in a container.
When velum is updated, it will be by launching a new container, thus the new
container should run db:migrate.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>