Bug 1031352 - Do not expose mariadb to the world
Switch to a UNIX socket for communication with MariaDB. The UNIX socket
is in a Kubernetes emptyDir so that it can be used among the other
containers in the pod.

Turned on set -e in the bash scripts, and stopped outputting the
db:setup to /dev/null as part of diagnosing startup problems.

You can't seem to use ConfigMaps in disconnected scenarios, so the
MariaDB configuration that disables networking is mounted from a
hostDir.

Using a hostPath to store the UNIX socket, since two separate pods need
access to it.

"public pod", which has hostNetwork: true contains:

* salt-master
* velum dashboard
* salt-api
* salt-minion-ca

"private pod", which has hostNetwork: false contains:

* velum-event-processor
* mariadb