Request the ca certificate directly to the CA machine (container)
The ca certificate was retrieved from the mine. The mine gets wiped
when the admin node gets rebooted (the salt-master container gets
rebooted). We don't want to wait for the mine to be populated if we
actually have this information in the filesystem.

If an orchestration is run, this information will be automatically
populated on the mine (but we don't want to force that), so we
directly use salt to cat the ca certificate from the ca machine (it's
a container at the moment).