Explicitly ask for the dashboard location
It will come with a reasonable default, but the person bootstrapping
the cluster will have the chance to set a different value so the
rest of the machines that will form part of the cluster will use this
reference during the bootstrapping process.

This modifies the workflow of the bootstrapping. To avoid problems, now
it's required for the customer/user to fill in all the first step
information in velum (registering first) for the autoyast profile to be
served (since it will contain a reference to the salt-master that will be
set on the rest of the machines). This way, the reference to the salt-master
on the autoyast profile, as well as the reference to the administration node
during the bootstrapping phase will be the one the customer/user set,
intead of being inferred from how the customer/user visited velum.

Until this first step of the setup is completed (dashboard reachable
location by the rest of the cluster for autoyast and salt) it won't be
possible to bring up new nodes (the autoyast profile will not be served, so
they will just raise an error in the YaST UI saying they can't retrieve
the autoyast configuration, since a 503 error will be returned).

Fixes bsc#1032461
Fixes bsc#1031605