Do not merge: Support selection of multiple masters
This change allows for the selection of multiple masters when bootstrapping
the cluster. The bulk of the changes to allow for this have been added in
the previous commit, allowing our diff from master to remain as small as is
possible. There is no immediate intent to merge this to master. As support
for multi master increases, we'll likely discard this change and work with
some UX folks to decide how best to display this.