Improve initial cluster config page [bsc#1040367]
This is not the definitive design of the initial cluster configuration
page, it's just an attempt to quickly improve it.

The different forms have been grouped together based on the area they
are related to.

Also hover help tips have been added to the forms that might be harder
to understand.

This helps to mitigate issue bsc#1040367

Signed-off-by: Flavio Castelli <fcastelli@suse.com>