When setting the highstate resolution, consider all highstates
With the current logic, it could happen that during the same
orchestration run the first highstate succeeds (setting the minion
status to :applied as succeeded), while the second highstate fails
(and the :applied status as succeeded will be kept).

With this change we consider the latest highstate run is the most
recent, and whatever it reports is what the final status of the
minion will be.