Check that result isn't `nil` before checking result contents
We are checking that the referenced file exists by performing a
rudimentary check. However, the result might be completely empty
under certain circumstances, making our check fail (`include?` called
on `nil` will fail).

This fixes the 500 error, but we still need to find out why the salt
side of things is failing at all to retrieve the contents of the
file.

Fixes: bsc#1043843