Added 'cluster is too small' warning
The etcd configuration used in CaaSP 1.0 requires
three total nodes to be present in the cluster
(can be 1 master and two minions) - without this,
etcd does not elect a leader, and bootstrap will fail.

Warns the user if he tries to setup a cluster with only
two nodes.

See bsc#1040477