Add support for running Puma with SSL
Puma will be executed using SSL. If the port provided is 80 (we will
be serving autoyast with no safe transport in 1.0), then we only
create the autoyast route, since we should not let the customers handle
Velum through an insecure connection.

Connect to the salt-api using TLS.

Remove certificate details from the UI and related code.

Fixes: bsc#1043570