Redirect from HTTP to HTTPS
This will not be applied to some paths like /_health or /autoyast.
Moreover, I haven't added tests for this because doing so is a can of
worms for Capybara (the existing tests), and it's not worth it.

Also note that the proper fix should come from a reverse-proxy side. We
need this for other features as well, but it would bring too much
complexity for the 1.0 release.

Fixes bsc#1047065

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>