Take orchestration result into consideration when bootstrapping
When bootstrapping, if the orchestration fails, depending in what stage
it failed, we might never retrieve individual highstate returns from minions,
so they will be in `pending` status forever.

With this change, we also listen for the overall bootstrap orchestration
result (`orch.kubernetes` for now), so we will mark all pending nodes as
failed.

Fixes: bsc#1046706