`force_ssl` is documented to be executed from the class context
Not all redirections to `https` were properly handled. By making this
change we manually redirect to `https` as a `before_action` as before,
but handling a `redirect_to` in the chain (effectively stopping the
chain and in the expected order).

The reason behind this is that `force_ssl` called from our own filter:
https://github.com/rails/rails/blob/03925dc26a747075ff37660d1f0a060a8178bf66/actionpack/lib/action_controller/metal/force_ssl.rb#L62-L68
installs a `before_action`, but it's already too late for that very
request (we are already in the `before_actions` for that request).

Also, with the current implementation the filter gets installed after
our other `before_action` filters, what implies that the redirection
to `https` is not the very first filter in the chain, as it should
be.

By calling to the method itself:
https://github.com/rails/rails/blob/03925dc26a747075ff37660d1f0a060a8178bf66/actionpack/lib/action_controller/metal/force_ssl.rb#L76-L96
from our filter, we ensure all requests are redirected to TLS.

Fixes: bsc#1048134