Detect orchestration results when `mods` has been provided as argument
We usually run the orchestrations manually as:
`salt-run state.orchestrate orch.kubernetes`

And this case was detected properly. This change adds support for also
detecting the orchestration result when the orchestration is called using
the `salt-api` on Velum (the general case):
`salt-run state.orchestrate mods=orch.kubernetes`

With this change we support both return results.

Fixes: bsc#1046706