Clean old jobs on our own, since salt won't do it automatically
Given our current configuration in which we are no longer using the
mysql `job_cache`, salt won't clean old jobs on its own, so we should
do it so the database won't keep growing without control.

We are executing raw SQL commands to avoid creating `Jid` and `SaltReturn`
models just for the purpose of cleaning, as we know that we want to move
away from a mysql backed salt master.

This is a follow up of: https://github.com/kubic-project/salt/pull/199

Fixes: bsc#1054256