ux: improved messages for 'Next' button on nodes selection
The title message for the 'Next' button on the nodes selection
page were out of date and weren't helping the user understand
what was wrong.

For each rule we use to validate the 'Next' button we
are adding a dedicated message to show to the user. This way
the user will know exactly what's wrong with the cluster
configuration.

I also fixed the "minimum of three nodes" alert to only be displayed
if the user didn't select 3 nodes (dones't matter if master was not
selected). The previous behavior was making it display if the
configuration wasn't supported even having 4 nodes selected. This
info now goes on the 'Next' button title message.

See bsc#1058915