Wait for orchestration result for marking success on minions
We are marking success depending on highstates, what makes a bit misleading
for the overall orchestration process. At this moment, orchestration will wait
for some tasks to have been completed (like some `kubectl apply` commands, which
apply immediately). We aren't waiting for the global orchestration result, and
this means that all minions might be marked as successful when we haven't really
waited for the overall orchestration to have finished.

This can have important implications, like not waiting for Dex to be completely
ready while the orchestration is waiting for it.

We just mark failed highstates as an early failure, but wait for the orchestration
result to finish and be succesful in order to mark pending nodes as successful.

This will also align with the current restritions in the UI to enable the `kubectl config`
button, since it will wait for the kubernetes master to be ready in order to be
enabled.

Fixes: bsc#1062542