When roles are set, set `not_applied` if it's a local assignment.
If we are setting the role locally only (on the database), set the
minion status to `not_applied`. If we are also setting the remote role
(set on salt too), set the minion status to `pending`.

Since there's no Orchestration model yet, there's no entity to control
this assignments, and Minion model will set one or the other depending
on the `remote` attribute of `assign_role`.

Fixes: bsc#1063977