Notify user when not able to fetch data from Velum API
We are regenerating Velum's certificate by adding the external
dashboard fqdn/ip to the Velum certificate. For this certificate
to be reloaded we are currently doing a `docker restart` on Velum
while the orchestration is being executed.

This has a bad effect on the user session as he/she is waiting
for the polling to finish. Without them noticing, the polling will
start failing as the certificate changed under them, and they need
to accept it, but ajax requests will just fail because of this.

With this patch we are notifying the user that the browser cannot
connect with Velum API and suggests he/she to try to reload the session.

Fixes bsc#1064641