Improved failure alert message
This is an attempt to make the experience better. We thought that
the previous message was too aggressive. Since this message will
always be shown for 2.0 at least, we want to tranquilize the user
that it's an expected behavior.

(cherry picked from commit 758d9f597f3453a8dd8a215d62b344f2cb548cd5)