Fix packaging by forcing the subset of gems to be installed.
Our `Gemfile` and `Gemfile.lock` should be the same under all environments,
development, test and production. Using an envvar on the Gemfile to decide
the dependency tree might cause deep problems and differences on the
dependency tree between development, test and production. We want to have
the same gem dependency tree on all possible environments, so we don't hit
bugs in production caused by a different dependency tree than the one used
in development and test, as this bugs could only be hit in production and
never reproduced in development and test environments.