skip bundler during spec file creation
a bundler BuildRequire is already hardcoded

other than that it prevents a false strict dependency on bundler, e.g.
if i run make_spec.sh on openSUSE 42.1 it will render bundler 1.10.6
whatever is installed on the system. this happened in our concourse
build container that uses 42.1 as a base system

Signed-off-by: Maximilian Meister <mmeister@suse.de>