Rename "Retry cluster upgrade" to "Retry cluster update".
I didn't refactor down the line because ultimately we cannot have an
enum with a :update status (rails will complain, so at least the enum
on the Orchestration model would need to be called :upgrade).

For that reason, keep everything as it is (call internally `upgrade`,
but just show "Retry cluster update" on the button label).