Improved CPI behavior with cloud frameworks
If a cloud framework is previously set, CPI value, if enabled, would be
based on what was set.

For example, if cloud:framework pillar value is "ec2", we would
set "ec2" as the CPI value if enabled. The same be applied for GCE and
Azure.

Otherwise, "openstack" would be the default value for CPI if enabled.

feature#cpi

Signed-off-by: Vítor Avelino <vavelino@suse.com>