ui: allowed master node to be added to the cluster
So far we only allowed workers to be added to the cluster through the
unassigned nodes page.

With this patch, we are now allowing the users to select the roles of the
nodes they want to add to the cluster. Very similar to what is available
from the discovery bootstrap page.

This feature keep forcing the user to respect the contrainst of having an
odd number of master nodes so we have a stable topology.

feature#add_master_node

Signed-off-by: Vítor Avelino <contact@vitoravelino.me>