Fixed # of masters warning when adding a new master
This patch fixes the warning that was being shown when the user selected
a new master node to be added and it wasn't breaking the odd number
constraint. Before this, the current assigned masters were not being
considered to show/not show the warning.

Now the warning will only be shown if the number of masters to be added
breaks the constraint as it's expected to be.

bsc#1087267

Signed-off-by: Vítor Avelino <contact@vitoravelino.me>