Avoid the event processor from crashing if it cannot interpret the event arguments
If we executed: `salt-run state.orchestrate 2` we get an error that crashes the event
processor:

`TypeError: no implicit conversion of String into Integer`

This happens because of `fun_args.first["mods"]` being `2["mods"]`, what makes it
crash with the previous exception. `2.respond_to?(:[])` is `true`, so I think the
best thing we can do here is to swallow the exception, and ignore the event if we
cannot even process its arguments.

Fixes: bsc#1088597