node removal: improved way of showing removal links
'Remove' and 'Force remove' links were being shown at the same time so
far. But 'Force remove' only makes sense to be shown when a previous
removal has failed.

With this we are showing only one link at a time. When the highstate of
a minion is fine, we show 'Remove'. Whenever it comes 'removal_failed',
we show 'Force remove'.

enhancemente#node-removal

Signed-off-by: Vítor Avelino <vavelino@suse.com>