Allow users to choose flannel's backend
Starting with v3 we changed the default flannel backend from UDP to
VXLAN.

The VXLAN backend does not work when users have segmented networks,
while the UDP one does.

The salt states are already capable of switching between the two
backends, we just needed to expose the right pillars from Velum.

This commit changes the setup UI to allow users to choose the backend of
and the port used by flannel.

feature#flannel-backend

Signed-off-by: Flavio Castelli <fcastelli@suse.com>