reservations: don't save if a value is not valid
Before this commit, the transaction on the `create` method from the
KubeletComputeResourcesReservationsController was useless because it
updated the attributes before the transaction itself. This commit
corrects that and simplifies it, so an error on any of the parameters
will issue a rollback (hence, the DB will be left untouched).

bsc#1096169

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>