ui: added placeholder for mirror registry blank option
The default blank option for the registry on the mirror creation form
was confusing. To keep the same experience with the "Create new
registry" button, the solution was to simply add "Select a registry"
text to work as a placeholder.

bsc#1095255

Signed-off-by: Vítor Avelino <vavelino@suse.com>
(cherry picked from commit 7065bbcd70901f58c4235542ea67462b2bc33769)