ui: do not count failed removal as minion for removal
We should not consider a failed removal minion as a valid removal when
we decide to show/hide the removals link.

In this case, if we are in a scenario 1+2 and one of the workers is in
removal failed highstate, we should not show the 'Remove' link for the
other worker.

bsc#1095205

Signed-off-by: Vítor Avelino <vavelino@suse.com>