Make 503 error message "branding agnostic"
This way this error message can be used both in kubic and in caasp

This is needed because we need to revert a change in the velum-branding
package and the caasp-container-manifests that was mounting the public
directory from the host.

Mounting the public directory from the host, was "hidding" the assets files
present in the public directory and so our images were broken.

This was only happening in the iso and not in our dev environments
because in production we have the assets precompiled and we don't
compile them on the fly, since we don't have a js engine installed.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>