Serve the flannel backend on the Pillar API
If we don't serve the flannel backend on the Pillar API, the upgrade
will use the default in the `sls` files, that is: `vxlan`.

The intent was always to keep `udp` for upgraded systems until we
migrate them to the `vxlan` backend, so make the Pillar controller
serve the current `flannel:backend` pillar value that is present
on the database (it will be `vxlan` for new installations, and `udp`
for systems coming from the 2.0 release).

Fixes: bsc#1096710