Expose pillars managing kubernetes extra args
The salt backend already has pillars that can be used to provide extra
arguments to the different kubernetes components (api-server,
controller-manager, scheduler, kubelet and proxy).

This commit allows to provide values for these settings. This is
something we should **not** expose into the UI. It can be handy with
certain customers who are looking for a way to pass not-yet-handled
flags to kubernetes.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

feature#hidden-kubernetes-pillars