ui: don't show failed alert for repeated orchestration
A failed alert is shown whenever one of the nodes fails to bootstrap.
This is an optimization to only show this alert once for a specific
orchestration run. If the user closes the alert, it won't come back
unless a new orchestration happens.

bsc#1097752

Signed-off-by: Vítor Avelino <vavelino@suse.com>
(cherry picked from commit e943e073e813cdea58bf0943e1caba226cb60131)