Updated tool tips and removed default data
 - Added a database migration that removed default data from the database
 - Added smart documentation on each LDAP field in place of the default data.

updates to the rspec tests too account for missing feilds due to the lack of defaults

removed not required element ids

removed not required element ids and corrected accidental removal of @certificate_holder.port

Removal of not needed fill_ins

Under a previous assumption of the workings of the save button I had filled in stuff that is not needed for the test.

Updated slim html to use Certificate

Also updated unit tests to pass.

Removal of extra line

Corrected spelling and fixed spacing

Corrected spelling and updated spacing to be consistent in the examples.

update to tests to fix a bug

There are some issues with these test causing random failures, Currently the best fix to this is locking down the run order.

fixed rubocop complaints

updated to fix rubocop complaints.