Fix rubocop issues
Fixed all of the various issues that rubocop complains about.
One of them was that two of the tests were duplicates. The test for no alt names wasn't actually testing against a cert without any altnames, so ended up having to alter the bash cert creation script to be able to properly generate a cert without altnames as that feature of the script was not working.